+++ src/containers/Router/index.js
@@ -1,15 +1,15 @@
 import React from "react";
 import { ConnectedRouter } from "connected-react-router";
 import { Switch, Route } from "react-router-dom";
-import LoginPage from "../Autenticacao/Login";
-import RegisterPage from "../Autenticacao/RegisterProfile";
+import LoginPage from "../Auth/Login";
+import RegisterPage from "../Auth/RegisterProfile";
 import Profile from "../ProfilePage";
 import ProfileEdit from '../ProfilePage/editProfile';
 import AddressEdit from '../ProfilePage/editAddress';
 import RestaurantPage from '../../components/Restaurant Page'
 import FeedRestaurants from "../../containers/FeedRestaurants";
 import CartPage from "../CartPage";
-import AddressRegister from "../Autenticacao/Address";
+import AddressRegister from "../Auth/Address";
 import FirstPage from "../FirstPage";
 import InputSearchResult from '../InputSearchResult/InputSearchResult'
